Adicionando o leitor csv

+++ src/br/com/acme/model/Drone.java
@@ -6,11 +6,11 @@ public class Drone {
 
 	private double cargaMaxima;
 
-	private int autonomiaKm;
+	private double autonomiaKm;
 
 	private Localizacao base;
 
-	public Drone(int identificador, double cargaMaxima, int autonomiaKm, Localizacao base) {
+	public Drone(int identificador, double cargaMaxima, double autonomiaKm, Localizacao base) {
 		this.identificador = identificador;
 		this.cargaMaxima = cargaMaxima;
 		this.autonomiaKm = autonomiaKm;
